Rename: 로그인 여부 모달 공통 컴포넌트로 이동
